+++ R/od-funs.R
@@ -21,6 +21,8 @@
 #' o1 = od_data_centroids2[od_data_centroids2[[1]] == od_data_df2[[1]][1], ]
 #' d1 = od_data_destinations[od_data_destinations[[1]] == od_data_df2[[2]][1], ]
 #' plot(desire_lines_d$geometry)
+#' plot(od_data_centroids2$geometry, add = TRUE, col = "green")
+#' plot(od_data_destinations$geometry, add = TRUE)
 #' plot(o1, add = TRUE)
 #' plot(d1, add = TRUE)
 #' plot(desire_lines_d$geometry[1], lwd = 3, add = TRUE)
@@ -97,6 +99,9 @@ od_to_sfc = function(x,
 #' pd = od_data_destinations
 #' od_coordinates(x, p, pd)
 od_coordinates = function(x, p = NULL, pd = NULL, silent = TRUE, sfnames = FALSE) {
+  if(methods::is(x, "sf")) {
+    return(od_coordinates_sf(x))
+  }
   o_code = x[[1]]
   d_code = x[[2]]
   if(methods::is(o_code, "factor")) {
@@ -150,6 +155,13 @@ od_coordinates = function(x, p = NULL, pd = NULL, silent = TRUE, sfnames = FALSE
   }
   odc = cbind(o_coords, d_coords)
   if(sfnames) return(as.matrix(odc)) # return without updating column names
+  colnames(odc) = c("ox", "oy", "dx", "dy")
+  odc
+}
+od_coordinates_sf = function(x) {
+  coords_o = sf::st_coordinates(lwgeom::st_startpoint(x))
+  coords_d = sf::st_coordinates(lwgeom::st_endpoint(x))
+  odc = cbind(coords_o, coords_d)
   colnames(odc) = c("ox", "oy", "dx", "dy")
   odc
 }