+++ NEWS.md
@@ -1,3 +1,8 @@
+# od 0.3.0 (2021-07)
+
+* New function `od_jitter()` allows offsetting of origin and destination points of desire lines (#11)
+* Add new argument to `od_disaggregate()` to allow non-integer outputs
+
 # od 0.2.1
 
 * Updated `od_disaggregate()` so it's less likely to fail with unfathomable message (#30)